+++ main.py
@@ -11,7 +11,7 @@ import tensorflow as tf
 
 from datapipeline.load_imageds import (  # model pipeline for loading image datasets
     LoadData, PredictionDataLoader)
-from models import EfficientNetB0Model  # models for training
+from models import EfficientNetB0Model, VanillaCNNModel  # models for training
 from models import (DenseNetModel, MobileNetModel, ResnetV2Model, VGG16Model,
                     XceptionNetModel)
 from trainer import ModelManager  # model manager for handing all the ops
@@ -22,7 +22,8 @@ MODEL_ARCH = {
     "vgg": VGG16Model,
     "efficientnet": EfficientNetB0Model,
     "resnet": ResnetV2Model,
-    "mobilenet": MobileNetModel
+    "mobilenet": MobileNetModel,
+    "vanilla":VanillaCNNModel
 }
 
 
@@ -117,7 +118,7 @@ if __name__ == "__main__":
     parser_train.add_argument("--model_arch",
                               choices=[
                                   "xception", "densenet", "efficientnet",
-                                  "vgg", "resnet", "mobilenet"
+                                  "vgg", "resnet", "mobilenet","vanilla"
                               ],
                               default="xception",
                               dest="model_arch")